menu updates

+++ upload/admin/model/tool/menu.php
@@ -26,7 +26,7 @@ class Menu extends \Opencart\System\Engine\Model {
 	 * $menu_id = $this->model_tool_menu->addMenu($code, $description, $cycle, $action, $status);
 	 */
 	public function addMenu(array $data): int {
-		$this->db->query("INSERT INTO `" . DB_PREFIX . "menu` SET `code` = '" . $this->db->escape($data['code']) . "', `route` = '" . $this->db->escape($data['route']) . "', `path` = '" . $this->db->escape($data['path']) . "', sort_order = '" . (int)$data['sort_order'] . "'");
+		$this->db->query("INSERT INTO `" . DB_PREFIX . "menu` SET `code` = '" . $this->db->escape($data['code']) . "', `type` = '" . $this->db->escape($data['type']) . "', `route` = '" . $this->db->escape($data['route']) . "', `path` = '" . $this->db->escape($data['path']) . "', sort_order = '" . (int)$data['sort_order'] . "'");
 
 		$menu_id = $this->db->getLastId();
 
@@ -39,22 +39,20 @@ class Menu extends \Opencart\System\Engine\Model {
 		return $menu_id;
 	}
 
-	public function editMenu(int $menu_id, array $data): int {
+	public function editMenu(int $menu_id, array $data): void {
 		$menu_info = $this->getMenu($menu_id);
 
-		$this->db->query("UPDATE `" . DB_PREFIX . "menu` SET `code` = '" . $this->db->escape($data['code']) . "', `route` = '" . $this->db->escape($data['route']) . "', `path` = '" . $this->db->escape($data['path']) . "', sort_order = '" . (int)$data['sort_order'] . "' WHERE `menu_id` = '" . (int)$menu_id . "'");
+		if ($menu_info) {
+			$this->db->query("UPDATE `" . DB_PREFIX . "menu` SET `code` = '" . $this->db->escape($data['code']) . "', `type` = '" . $this->db->escape($data['type']) . "', `route` = '" . $this->db->escape($data['route']) . "', `path` = '" . $this->db->escape($data['path']) . "', sort_order = '" . (int)$data['sort_order'] . "' WHERE `menu_id` = '" . (int)$menu_id . "'");
 
-		if ($menu_info['path'] != $data['path']) {
 			$this->db->query("UPDATE `" . DB_PREFIX . "menu` SET `path` = REPLACE(`path`, '" . $this->db->escape($menu_info['path']) . "_', '" . $this->db->escape($data['path']) . "_') WHERE `path` LIKE '" . $this->db->escape($menu_info['path']) . "_%'");
+
+			$this->model_tool_menu->deleteDescriptions($menu_id);
+
+			foreach ($data['menu_description'] as $language_id => $menu_description) {
+				$this->model_tool_menu->addDescription($menu_id, $language_id, $menu_description);
+			}
 		}
-
-		$this->model_tool_menu->deleteDescriptions($menu_id);
-
-		foreach ($data['menu_description'] as $language_id => $menu_description) {
-			$this->model_tool_menu->addDescription($menu_id, $language_id, $menu_description);
-		}
-
-		return $menu_id;
 	}
 
 	/**
@@ -71,9 +69,15 @@ class Menu extends \Opencart\System\Engine\Model {
 	 * $this->model_tool_menu->deleteMenuByCode($code);
 	 */
 	public function deleteMenu(string $menu_id): void {
-		$this->db->query("DELETE FROM `" . DB_PREFIX . "menu` WHERE `menu_id` = '" . (int)$menu_id . "'");
+		$menu_info = $this->getMenu($menu_id);
 
-		$this->model_tool_menu->deleteDescriptions($menu_id);
+		if ($menu_info) {
+			$this->db->query("DELETE FROM `" . DB_PREFIX . "menu` WHERE `menu_id` = '" . (int)$menu_id . "'");
+
+			$this->db->query("UPDATE `" . DB_PREFIX . "menu` SET `path` = REPLACE(`path`, '" . $this->db->escape($menu_info['path']) . "', '" . $this->db->escape(substr($menu_info['path'], 0, strrpos($menu_info['path'], '_'))) . "') WHERE `path` LIKE '" . $this->db->escape($menu_info['path']) . "_%'");
+
+			$this->model_tool_menu->deleteDescriptions($menu_id);
+		}
 	}
 
 	/**
@@ -90,10 +94,10 @@ class Menu extends \Opencart\System\Engine\Model {
 	 * $this->model_tool_menu->deleteMenuByCode($code);
 	 */
 	public function deleteMenuByCode(string $code): void {
-		$menu_info = $this->getMenuByCode($code);
+		$results = $this->getMenuByCode($code);
 
-		if ($menu_info) {
-			$this->model_tool_menu->deleteMenu($menu_info['menu_id']);
+		foreach ($results as $result) {
+			$this->model_tool_menu->deleteMenu($result['menu_id']);
 		}
 	}
 
@@ -119,25 +123,6 @@ class Menu extends \Opencart\System\Engine\Model {
 	}
 
 	/**
-	 * Get Menu By Code
-	 *
-	 * @param string $code
-	 *
-	 * @return array<string, mixed>
-	 *
-	 * @example
-	 *
-	 * $this->load->model('tool/menu');
-	 *
-	 * $menu_info = $this->model_tool_menu->getMenuByCode($code);
-	 */
-	public function getMenuByCode(string $code): array {
-		$query = $this->db->query("SELECT DISTINCT * FROM `" . DB_PREFIX . "menu` `m` LEFT JOIN `" . DB_PREFIX . "menu_description` `md` ON (`m`.`menu_id` = `md`.`menu_id`) WHERE `m`.`code` = '" . $this->db->escape($code) . "' AND `md`.`language_id` = '" . (int)$this->config->get('config_language_id') . "'");
-
-		return $query->row;
-	}
-
-	/**
 	 * Get Menu(s)
 	 *
 	 * Get the record of the menu records in the database.
@@ -152,8 +137,16 @@ class Menu extends \Opencart\System\Engine\Model {
 	 *
 	 * $results = $this->model_tool_menu->getMenus();
 	 */
-	public function getMenus(): array {
-		$query = $this->db->query("SELECT * FROM `" . DB_PREFIX . "menu` `m` LEFT JOIN `" . DB_PREFIX . "menu_description` `md` ON (`m`.`menu_id` = `md`.`menu_id`) ORDER BY `m`.`path` ASC");
+	public function getMenus($path = ''): array {
+		$sql = "SELECT * FROM `" . DB_PREFIX . "menu` `m` LEFT JOIN `" . DB_PREFIX . "menu_description` `md` ON (`m`.`menu_id` = `md`.`menu_id`)";
+
+		if ($path) {
+			$sql .= " WHERE `m`.`path` LIKE '" . $this->db->escape($path) . "'";
+		}
+
+		$sql .= " ORDER BY `m`.`path` ASC, `m`.`sort_order` ASC";
+
+		$query = $this->db->query($sql);
 
 		return $query->rows;
 	}
